refactor(folders) inputs and outputs, not i/o_plugins
