+++ livelossplot/inputs/pytorch_ignite.py
