+++ livelossplot/outputs/base_output.py
