+++ livelossplot/outputs/extrema_print.py
