+++ livelossplot/outputs/matplotlib_subplots.py
